data_warehouse/requirements.txt
mysql-connector-python
pandas
pyarrow
matplotlib